app/cloud-sql-exporter/requirements.txt
google-api-python-client==1.7.11
google-auth==1.11.2
SQLAlchemy==1.3.13
PyMySQL==0.9.3
